fix: переносит функцию toggleFullScreen из компоненты в utils
